Command-line tool output array dictionary
Also refactored PathEplorers and added a license
