+++ Sources/ScoutCLT/ReadCommand.swift
@@ -83,11 +83,14 @@ struct ReadCommand: ParsableCommand {
         var value: String
 
         if let json = try? PathExplorerFactory.make(Json.self, from: data) {
-            value = try json.get(readingPath).stringValue
+            let key = try json.get(readingPath)
+            value = key.string ?? key.description
         } else if let plist = try? PathExplorerFactory.make(Plist.self, from: data) {
-            value = try plist.get(readingPath).stringValue
+            let key = try plist.get(readingPath)
+            value = key.string ?? key.description
         } else if let xml = try? PathExplorerFactory.make(Xml.self, from: data) {
-            value = try xml.get(readingPath).stringValue
+            let key = try xml.get(readingPath)
+            value = key.string ?? key.description
         } else {
             if let filePath = inputFilePath {
                 throw RuntimeError.unknownFormat("The format of the file at \(filePath) is not recognized")
